Weather Forecast

Get weather predictions for Austria/Alps

Usage

geosphere_get_prediction(
  host,
  x,
  y,
  level = NULL,
  subdatasets = NULL,
  download_dir = "data_raw",
  timezone = "Europe/Paris"
)

Arguments

host

Character. The data source to query. Must be one of:

"nwp"

Numerical Weather Prediction (NWP) model outputs. These are high-resolution forecasts for the extended Alpine region on a 2.5 km grid, updated every 3 hours with a forecast horizon of 60 hours. For more information, visit https://data.hub.geosphere.at/dataset/nwp-v1-1h-2500m.

"nowcast"

Short-term forecasts using real-time observations and radar data, useful for rapid updates. For more information, visit https://data.hub.geosphere.at/dataset/nowcast-v1-15min-1km.

"ensemble"

The C-LAEF (Convection-permitting Limited-Area Ensemble Forecasting) system provides weather forecasts for the extended Alpine region on a 2.5 km grid. It is updated twice daily with a forecast horizon of 60 hours. The dataset includes the 10th, 50th, and 90th percentiles from 17 model runs (16 ensemble members + 1 control run), allowing for uncertainty estimation in the forecasts. For more information, visit https://data.hub.geosphere.at/dataset/ensemble-v1-1h-2500m.

x

the coordinates of the point to get the prediction for

y

the coordinates of the point to get the prediction for

level

the level of the prediction to get

subdatasets

the subdatasets to get

download_dir

the directory to download the data to

timezone

the timezone of the prediction

Value

a data.frame with the prediction